#16 Move TokenManager from util to format_prediction
